+++ charitable/includes/elementor/widgets/class-charitable-elementor-button-widget.php
@@ -215,7 +215,7 @@ class Charitable_Elementor_Button_Widget extends \Elementor\Widget_Base
 
         $output .= '</div>';
 
-        echo wp_kses_post( $output );
+        echo $output; // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
     }